tests: move init logging tests to its own module
