+++ src/autopatchshared/autopatchshared/test_init_logging.py
@@ -1,0 +1,68 @@
+import json
+import logging
+import logging.config
+import os
+from unittest import mock
+from unittest.mock import mock_open
+from .init_logging import init_logging
+
+# --------------
+# Tests for init_logging
+# --------------
+
+
+def test_init_logging_file_not_found(monkeypatch, capsys):
+    # Force os.path.exists to always return False
+    monkeypatch.setattr(os.path, "exists", lambda path: False)
+    # Call init_logging with a dummy path.
+    logger = init_logging("dummy_logging.json", "test_app")
+    captured = capsys.readouterr().out
+    assert "not found" in captured
+    assert "Falling back to basic logging configuration." in captured
+    # Even in fallback, a logger with the given name is returned.
+    assert logger.name == "test_app"
+
+
+def test_init_logging_invalid_json(monkeypatch, capsys):
+    # Simulate that the logging config file exists.
+    monkeypatch.setattr(os.path, "exists", lambda path: True)
+    # Provide invalid JSON content.
+    m = mock_open(read_data="not valid json")
+    monkeypatch.setattr("builtins.open", m)
+    logger = init_logging("dummy_logging.json", "test_app")
+    captured = capsys.readouterr().out
+    assert "Error decoding JSON" in captured or "Unexpected error" in captured
+    assert "Falling back to basic logging configuration." in captured
+    assert logger.name == "test_app"
+
+
+def test_init_logging_valid(monkeypatch, capsys):
+    # Provide a valid logging configuration.
+    valid_config = {
+        "version": 1,
+        "formatters": {"default": {"format": "%(levelname)s:%(message)s"}},
+        "handlers": {
+            "console": {
+                "class": "logging.StreamHandler",
+                "formatter": "default",
+                "stream": "ext://sys.stdout",
+            }
+        },
+        "loggers": {
+            "test_app": {
+                "handlers": ["console"],
+                "level": "DEBUG",
+                "propagate": False,
+            }
+        },
+    }
+    config_str = json.dumps(valid_config)
+    # Simulate file exists.
+    monkeypatch.setattr(os.path, "exists", lambda path: True)
+    m = mock_open(read_data=config_str)
+    monkeypatch.setattr("builtins.open", m)
+    logger = init_logging("dummy_logging.json", "test_app")
+    # The logger should be configured and have logged the initialization.
+    captured = capsys.readouterr().out
+    assert "Logger initialized successfully." in captured
+    assert logger.name == "test_app"
