+++ src/fuzzing-service/test_fuzzing_service.py
@@ -19,7 +19,6 @@ from fuzzing_service import (
     MapCrashDetailsAsCloudEvents,
     compile_program_run_fuzzer,
     extract_crashes,
-    init_logging,
     load_config,
     produce_output,
     write_crashes_csv,
@@ -153,68 +152,6 @@ async def test_produce_output_with_multiple_events(dummy_logger):
     ]
     await produce_output(crash_details)
     assert "Producing 2 CloudEvents." in dummy_logger.messages
-
-
-# --------------
-# Tests for init_logging
-# --------------
-
-
-def test_init_logging_file_not_found(monkeypatch, capsys):
-    # Force os.path.exists to always return False
-    monkeypatch.setattr(os.path, "exists", lambda path: False)
-    # Call init_logging with a dummy path.
-    logger = init_logging("dummy_logging.json", "test_app")
-    captured = capsys.readouterr().out
-    assert "not found" in captured
-    assert "Falling back to basic logging configuration." in captured
-    # Even in fallback, a logger with the given name is returned.
-    assert logger.name == "test_app"
-
-
-def test_init_logging_invalid_json(monkeypatch, capsys):
-    # Simulate that the logging config file exists.
-    monkeypatch.setattr(os.path, "exists", lambda path: True)
-    # Provide invalid JSON content.
-    m = mock_open(read_data="not valid json")
-    monkeypatch.setattr("builtins.open", m)
-    logger = init_logging("dummy_logging.json", "test_app")
-    captured = capsys.readouterr().out
-    assert "Error decoding JSON" in captured or "Unexpected error" in captured
-    assert "Falling back to basic logging configuration." in captured
-    assert logger.name == "test_app"
-
-
-def test_init_logging_valid(monkeypatch, capsys):
-    # Provide a valid logging configuration.
-    valid_config = {
-        "version": 1,
-        "formatters": {"default": {"format": "%(levelname)s:%(message)s"}},
-        "handlers": {
-            "console": {
-                "class": "logging.StreamHandler",
-                "formatter": "default",
-                "stream": "ext://sys.stdout",
-            }
-        },
-        "loggers": {
-            "test_app": {
-                "handlers": ["console"],
-                "level": "DEBUG",
-                "propagate": False,
-            }
-        },
-    }
-    config_str = json.dumps(valid_config)
-    # Simulate file exists.
-    monkeypatch.setattr(os.path, "exists", lambda path: True)
-    m = mock_open(read_data=config_str)
-    monkeypatch.setattr("builtins.open", m)
-    logger = init_logging("dummy_logging.json", "test_app")
-    # The logger should be configured and have logged the initialization.
-    captured = capsys.readouterr().out
-    assert "Logger initialized successfully." in captured
-    assert logger.name == "test_app"
 
 
 # --------------
